+++ beams/beams/doctype/compensation_proposal/compensation_proposal.js
@@ -1,7 +1,7 @@
 // Copyright (c) 2024, efeone and contributors
 // For license information, please see license.txt
 
-frappe.ui.form.on('Job Proposal', {
+frappe.ui.form.on('Compensation Proposal', {
     refresh: function(frm) {
         frm.set_query('job_applicant', function() {
             return {
