+++ beams/beams/doctype/compensation_proposal/compensation_proposal.json
@@ -17,7 +17,15 @@
   "section_break_jt",
   "job_offer_term_template",
   "column_break_ogbe",
-  "terms_and_conditions"
+  "terms_and_conditions",
+  "section_break_vcik",
+  "remarksfeedback",
+  "payslip_attachments_section",
+  "payslip_month_1",
+  "column_break_ctnw",
+  "payslip_month_2",
+  "column_break_vaog",
+  "payslip_month_3"
  ],
  "fields": [
   {
@@ -69,7 +77,7 @@
    "fieldtype": "Link",
    "label": "Amended From",
    "no_copy": 1,
-   "options": "Job Proposal",
+   "options": "Compensation Proposal",
    "print_hide": 1,
    "read_only": 1,
    "search_index": 1
@@ -93,15 +101,55 @@
    "fieldtype": "Link",
    "label": "Select Terms and Conditions",
    "options": "Terms and Conditions"
+  },
+  {
+   "fieldname": "section_break_vcik",
+   "fieldtype": "Section Break"
+  },
+  {
+   "fieldname": "remarksfeedback",
+   "fieldtype": "Small Text",
+   "label": "Remarks/Feedback"
+  },
+  {
+   "fieldname": "payslip_attachments_section",
+   "fieldtype": "Section Break"
+  },
+  {
+   "fieldname": "payslip_month_1",
+   "fieldtype": "Attach",
+   "label": "Payslip - Month 1",
+   "read_only": 1
+  },
+  {
+   "fieldname": "column_break_ctnw",
+   "fieldtype": "Column Break"
+  },
+  {
+   "fieldname": "payslip_month_2",
+   "fieldtype": "Attach",
+   "label": "Payslip - Month 2",
+   "read_only": 1
+  },
+  {
+   "fieldname": "column_break_vaog",
+   "fieldtype": "Column Break"
+  },
+  {
+   "fieldname": "payslip_month_3",
+   "fieldtype": "Attach",
+   "label": "Payslip - Month 3",
+   "read_only": 1
   }
  ],
+ "grid_page_length": 50,
  "index_web_pages_for_search": 1,
  "is_submittable": 1,
  "links": [],
- "modified": "2024-11-07 12:14:17.566123",
+ "modified": "2025-07-24 17:24:49.165221",
  "modified_by": "Administrator",
  "module": "BEAMS",
- "name": "Job Proposal",
+ "name": "Compensation Proposal",
  "naming_rule": "Expression (old style)",
  "owner": "Administrator",
  "permissions": [
@@ -147,6 +195,7 @@
    "write": 1
   }
  ],
+ "row_format": "Dynamic",
  "sort_field": "modified",
  "sort_order": "DESC",
  "states": [],
