+++ beams/beams/doctype/compensation_proposal/test_compensation_proposal.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestJobProposal(FrappeTestCase):
+class TestCompensationProposal(FrappeTestCase):
 	pass
